requirements.txt
flask==3.0.0
flask-login==0.6.3
flask-wtf==1.2.1
supabase==2.4.0
python-dotenv==1.0.0
uvicorn[standard]==0.30.0
fastapi==0.115.0
WTForms==3.0.1
Jinja2==3.1.2
